Use hook to only initiate the calls
