+++ core-ui/src/hooks/useFeature.ts
@@ -1,10 +1,10 @@
 import { useRecoilValue } from 'recoil';
-import { configurationState } from 'state/configurationSelector';
+import { configurationAtom } from 'state/configurationAtom';
 import { ConfigFeature, ConfigFeaturesNames } from 'state/types';
 
 export function useFeature<T extends ConfigFeature>(
   featureName: ConfigFeaturesNames,
 ): T | undefined {
-  const configuration = useRecoilValue(configurationState);
+  const configuration = useRecoilValue(configurationAtom);
   return configuration?.features?.[featureName] as T;
 }
